add team and team spec files
